requirements.txt
Flask==3.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
matplotlib==3.8.2
numpy==1.26.4
uuid==1.30
pytest==8.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
# psycopg2==2.9.9

